Move ResponsableGroupe controller
