+++ Tools/Controllers/AuthentificationController.php
@@ -27,13 +27,6 @@ implements Interfaces\IGetable
     /**
      * {@inheritDoc}
      */
-    protected function ensureAccessUser(string $order, \LibertAPI\Utilisateur\UtilisateurEntite $utilisateur)
-    {
-    }
-
-    /**
-     * {@inheritDoc}
-     */
     public function get(IRequest $request, IResponse $response, array $arguments) : IResponse
     {
         $authentificationType = 'Basic';
