+++ Tools/Route/Utilisateur.php
@@ -1,4 +1,7 @@
 <?php declare(strict_types = 1);
+
+use LibertAPI\Tools\Controllers\UtilisateurController;
+
 /*
  * Doit être importé après la création de $app. Ne créé rien.
  *
@@ -9,9 +12,9 @@
 $app->group('/utilisateur', function () {
     $this->group('/{utilisateurId:[0-9]+}', function () {
         /* Detail */
-        $this->get('', 'controller:get')->setName('getUtilisateurDetail');
+        $this->get('', [UtilisateurController::class, 'get'])->setName('getUtilisateurDetail');
     });
 
     /* Collection */
-    $this->get('', 'controller:get')->setName('getUtilisateurListe');
+    $this->get('', [UtilisateurController::class, 'get'])->setName('getUtilisateurListe');
 });
